Inline symbolize::cfg functionality directly into symbolize module
Move all contents of symbolize::cfg::* directly into the symbolize
module. That is done for symmetry with what we have for the inspect
functionality.
