+++ benches/gsym.rs
@@ -1,6 +1,6 @@
 use std::path::Path;
 
-use blazesym::symbolize::cfg;
+use blazesym::symbolize::Gsym;
 use blazesym::symbolize::Source;
 use blazesym::symbolize::Symbolizer;
 use blazesym::symbolize::SymbolizerFeature;
@@ -18,7 +18,7 @@ fn symbolize_end_to_end() {
         SymbolizerFeature::DebugInfoSymbols(true),
         SymbolizerFeature::LineNumberInfo(true),
     ];
-    let src = Source::Gsym(cfg::Gsym {
+    let src = Source::Gsym(Gsym {
         file_name: gsym_vmlinux,
         base_address: 0,
     });
