+++ src/lib.rs
@@ -9,15 +9,15 @@
 //!
 //! Here an example illustrating usage of the symbolization functionality:
 //! ```no_run
-//! use blazesym::symbolize::cfg;
 //! use blazesym::Addr;
 //! use blazesym::symbolize::Symbolizer;
 //! use blazesym::symbolize::Source;
+//! use blazesym::symbolize::Process;
 //! use blazesym::symbolize::SymbolizedResult;
 //!
 //! let process_id: u32 = std::process::id(); // <some process id>
 //! // Load all symbols of loaded files of the given process.
-//! let src = Source::Process(cfg::Process { pid: process_id.into() });
+//! let src = Source::Process(Process { pid: process_id.into() });
 //! let symbolizer = Symbolizer::new().unwrap();
 //!
 //! let stack: [Addr; 2] = [0xff023, 0x17ff93b];  // Addresses of instructions
