+++ src/symbolize/mod.rs
@@ -1,0 +1,20 @@
+mod source;
+mod symbolizer;
+
+use std::path::PathBuf;
+
+pub use source::Elf;
+pub use source::Gsym;
+pub use source::Kernel;
+pub use source::Process;
+pub use source::Source;
+pub use symbolizer::SymbolizedResult;
+pub use symbolizer::Symbolizer;
+pub use symbolizer::SymbolizerFeature;
+
+
+pub(crate) struct AddrLineInfo {
+    pub path: PathBuf,
+    pub line: usize,
+    pub column: usize,
+}
